scripts/training/requirements.txt
# API Training Script Dependencies

# HTTP requests with retry support
requests>=2.31.0

# Command-line interface framework
click>=8.1.7

# Environment variable management
python-dotenv>=1.0.0

# PostgreSQL database adapter
psycopg2-binary>=2.9.9

# Progress bars for CLI
tqdm>=4.66.1

# Testing dependencies (optional)
pytest>=7.4.3
pytest-mock>=3.12.0
